Add optimum Intel® pass for converting a Huggingface Model to an OpenVINO Model (#1759)
## Describe your changes

The changes in this PR are for creating an optimum Intel® pass for
converting a Huggingface Model to an OpenVINO model, along with the
ability to perform weight compression or quantization if necessary.
The aim is to add in the functionality offered by the [optimum-cli
export
openvino](https://huggingface.co/docs/optimum/main/en/intel/openvino/export)
command within Olive.

The changes include the addition of a new pass, unit tests for the pass,
a few updates to the other OpenVINO passes, and updates to the
documentation files as well.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [x] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link