Refactor OnnxOpVersionConversion to use the onnxscript version converter (#1784)
Refactor OnnxOpVersionConversion to use the onnxscript version converter
with the fallback option. When the version converter implemented in ONNX
IR supports the target version, the converter will be used. Otherwise,
we fall back to use the version converter in onnx. The API call handles
initializers to ensure models >2gb can be supported.