Add initial selective mixed precision pass (#1898)
## Describe your changes
Initial implementation of the selective mixed precision class.
- More algorithms and heuristics will be added later. Experiments for
data driven selection is underway.
- Two algorithms introduced. `k_quant_last` and `k_quant_mixed` similar
to those in #1818. However, qkv_proj is not added since it is a large
matrix and expensive. We need more testing to see if it's required to
keep qkv in higher precision and/or if we should split qkv.
- Downstream passes (both pytorch and onnx) should consume this
information to do mixed quantization. Some details regarding how to deal
with fused qkv needs to be sorted out. This pass is agnostic to such
downstream changes.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link