Phi3_5: Remove add metadata pass (#1922)
## Describe your changes
`AddOliveMetadata` does not work for static-llm models which have a
special structure. Removing from the NPU recipes for now since these are
important recipes.
The metadata could be added before the static-llm pass and the later
passes be updated to preserve the metadata in the components.

@ashrit-ms
## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link