Fix bug where dml provider was being changed to cpu
A set of if/elif/else statements was neglecting to account for the case of using dml, and was instead changing things to cpu in the final else block when the user tried using dml.  This commit changes the final else block to be elif provider == "cpu".  That way, if the provider is neither cuda, nor qnn, nor cpu, the last possibility is dml (which should be the default according to an earlier comment in line 326), and no changes are made.