Avoid running ByteSize() if save_as_external_data is enabled
ByteSize() used to be quick but capable of wildly wrong results with Protobuf 3.x.  Now with newer Protobuf versions, it seems more consistent, but is very costly to run, to the tune of ~60GB RAM/swap and >5 minutes for a ~10GB unet file (YMMV).  Therefore, change the code here to only run ByteSize() if save_as_external_data is not already enabled.