Exclude onnx and onnx.data files when copying unoptimized model's folder tree
Currently, the folder tree for the optimized model is created by copying the unoptimized model's folder tree, then overwriting the unoptimized files with optimized versions from the run cache.  Weights.pb files are ignored/excluded during the tree copy, which is fine.  However, there is a scenario (that actually occurs with the default SDXL 1.0 model used by the example) where text_encoder_2 can be >2GB when unoptimized, but <2GB when optimized.  In this case, the unoptimized model will have a model.onnx.data external data file, whereas the optimized model will not.  The unoptimized external file gets copied over to the optimized tree, but is not overwritten when the optimized model files are copied in, since there isn't an optimized version of it.  You end up with a vestigial 3GB file that isn't actually used by the optimized model.  Therefore, this commit changes the tree-copy code to also ignore model.onnx and model.onnx.data files (in addition to the currently-ignored weights.pb files).  The unoptimized onnx files don't actually need to be copied over, since they are overwritten anyway when the optimized models are copied in.