Move interactive GUI controls from bottom to top so they stay onscreen
Previously, the interactive GUI controls were at the bottom of the window.  But this tends to put them offscreen if there is more than one row of images, and they cannot be brought back into view by click-dragging the window titlebar, since the mouse pointer cannot drag higher than the top of the screen.  Maximizing the window isn't possible either, since the maximize button is disabled.
Therefore, this commit moves the GUI controls to the top of the window so they can remain reachable.  (This includes the text prompt field, the "Generate" button, and the progress bar.)