In interactive GUI, enable pressing "Enter" key in text field to start generation, besides clicking "Generate" button
Previously, in the interactive GUI, pressing "Enter" in the text field did nothing.  This commit enables pressing "Enter" in the text field to start generating, besides the existing way of clicking the "Generate" button.  This can help in particular if/when the window is wider than the screen, where the "Generate" button may be offscreen.