Add optional --image_width and --image_height flags
If the user does not specify values for these flags, they are set to the value of --image_size, which defaults to 768.  Internal functions now take width and height as params instead of size.   I also updated the help text to note that size/width/height values must be multiples of 32 (otherwise an error occurs).