Register EP libraries in ORT (#2052)
## Describe your changes
- New ORT design has EPs as libraries that can be registered on the core
runtime.
- New API is used only for Windows + ORT 1.23+. Linux support is not
well defined and the apis don't seem to work.
- Provide a way for users to provide the EP as tuple where the second
element is the local path to the library. This is optional and the ep
can still be provided as a single string.
- If the ORT package was built with additional eps such as cuda, vitisai
or qnn, we automatically register them using the name of the dll which
will be in the `onnxruntime/capi` directories with the other ort
libaries.
- Windows ML case is not supported yet since it's not clear what is
required.


## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link