Move docs build from azure pipeline to github actions (#2100)
## Move docs build from azure pipeline to github actions

The workflow is used only for verification purpose and not for
publish/release anymore. Release workflow is now in
Olive-Release-Workflows repository as it is required to be using
governed templates.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link