Quantization: lm_head quantization, k_quant_down heuristic, No zero offset while packing (#2137)
## Describe your changes
- `Gptq` and olive quant checkpoint now supports lm_head quantization
- `SelectiveMixedPrecision` has a new heuristic called `k_quant_down`
which is similar to `k_quant_mixed` but only takes the `down_proj`
linears and `lm_head`.
- `QuantLinear` packing now doesn't do the gptq style -1 offset to the
zero points while packing. This causes issues with asymmetric
quantization when any zero point value is 0 as the values are expected
to be unsigned int. It also now has a `from_tensor` method that can be
used to create a new instance from already quantized weights.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link