Add Shape to ALLOWED_CONSUMER_OPS in ReplaceAttentionMaskValue surgery (#2213)
## Describe your changes
In [text_encoder](https://huggingface.co/openai/clip-vit-large-patch14)
model, a ConstantOfShape op with -inf value is not getting replaced with
the given replacement value in ReplaceAttentionMaskValue surgery as
Shape node is one of the consumers of the ConstantOfShape node. Hence
added Shape to ALLOWED_CONSUMER_OPS.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link