Update extract adapter pass to match torch dynamo export (#2185)
## Describe your changes

We will only focus on cases when the onnx model is exported from torch
dynamo. If the optimize is True (which is default) for torch dynamo
export, the lora weights will be folded. So:

1. we check where the initializer comes from for each initializer.
2. If it is from a LORA weight, rename it back to lora weight, and save
it.
3. DORA weights might be folded to GEMM, we decompose this node to
MatMul so it can be quantized by Olive quantization pass.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link