Static Quantization: Patch `MinMaxCalibrator`, Handle `extra_options` for QNN EP (#2229)
## Describe your changes
- The MixMax calibrator stores all the outputs of each session run as
intermediate outputs. However, it only needs the augmented min and max
values. This causes the memory usage to unnecessarily grow with each
iteration of the calibration data, especially for llms which have large
outputs (logits and kv cache). This has been addressed in ort with
https://github.com/microsoft/onnxruntime/pull/26321
- Patching in Olive also with a more general solution by removing the
original outputs from the augmented graph as well as unused leaf nodes
to reduce memory and compute usage.
- The static quantization pass did not handle `extra_options` when the
EP is QNN. Fixing this as well.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link