LLMAugmentedDataLoader: No decode phase for non-GQA model (#2204)
## Describe your changes
The LLMAugmentedDataLoader had a decode phase where we ran the model to
get the past kv cache inputs. This was done because I thought this would
give the past inputs the same stats as the present outputs. However,
this is not the case since the inputs and outputs for the two phases are
different.
In my experiments, I found that it is easier to just run the prefill
phase during activation quantization and then use some graph surgeries
later to sync the quantization parameters between the past inputs and
present outputs.

So, we remove the decode phase in the dataloader.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link