TieWordEmbeddings: unquantized and quantized support (#2240)
## Describe your changes
- New surgery `TieWordEmbeddings` that ties the weights between the
input embeddings and lm head
- Two cases are supported:
- Both are unquantized: lm head MatMul is replaced with a Gemm with
transB set to True. input activation is reshaped to 2D and the output is
reshaped back to original rank. For CPU and CUDA atleast, reshape on
activation is just a metadata change so there is no overhead. This
option was chosen over adding a transpose node on the shared weight
since ORT constant folds the transpose and duplicates the initializer
during session init.
- Both are quantized with MatMulNBits and GatherBlockQuantized. A
reshape node is added on the MatMulNBits qweight. This only saves disk
space but not device memory because ORT constant folds the reshape and
duplicates the initializer. I plan to follow up with the contrib op
developers if we could relax the 2d shape requirement on the
GatherBlockQuantized op. Although I don't know if sharing the
initializers in memory would be useful for the CPU EP case since the
MatMulNBits op does kernel prepacking anyways.
 
## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link