SelectiveMixedPrecision: Sensitivity score based algorithms (#2265)
## Describe your changes
- New selective mixed precision algorithms added that select the layers
to put in higher precision based on sensitivity scores and a ratio.
- The sensitivity scores are signal-to-noise ratio, integer quantization
error and kl divergence gradient. Refer to the pass docstring for more
information on the scores.
- Some refactoring to reuse the default quantization calibration data
between the passes.

Average improvement of the various algorithms across different
(block_size, symmetric) settings compared with the full int4 model.
- Ratios were selected to match the model size of the score-based
algorithms to that of the `k_quant_down` heuristic.
- `kld_gradient` often performs among the top algorithms. However, it
requires GPU resources for efficient computation because it is
data-dependent and requires both forward and backward passes.

<img width="550" height="859" alt="image"
src="https://github.com/user-attachments/assets/eb08d195-daba-47f5-b7cd-74f8dd68687f"
/>



## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link